Move gui_info.py into RimSort.utils
